refactored notification filters
