+++ source/php/App.php
@@ -9,14 +9,15 @@ use VolunteerManager\Notification\NotificationHandler;
 use VolunteerManager\Notification\NotificationsConfig;
 use VolunteerManager\PostType\Application\Application;
 use VolunteerManager\PostType\Application\ApplicationConfiguration;
+use VolunteerManager\PostType\Application\ApplicationNotificationFilters;
 use VolunteerManager\PostType\Assignment\Assignment;
 use VolunteerManager\PostType\Assignment\AssignmentConfiguration;
-use VolunteerManager\PostType\Assignment\AssignmentNotifications;
+use VolunteerManager\PostType\Assignment\AssignmentNotificationFilters;
 use VolunteerManager\PostType\Employee\Employee;
 use VolunteerManager\PostType\Employee\EmployeeApiManager;
 use VolunteerManager\PostType\Employee\EmployeeApiValidator;
 use VolunteerManager\PostType\Employee\EmployeeConfiguration;
-use VolunteerManager\PostType\Employee\EmployeeNotifications;
+use VolunteerManager\PostType\Employee\EmployeeNotificationFilters;
 
 class App
 {
@@ -40,12 +41,12 @@ class App
         //Post types
         $assignment = new Assignment(...array_values(AssignmentConfiguration::getPostTypeArgs()));
         $assignment->addHooks();
-        $assignmentNotifications = new AssignmentNotifications();
+        $assignmentNotifications = new AssignmentNotificationFilters();
         $assignmentNotifications->addHooks();
 
         $employee = new Employee(...array_values(EmployeeConfiguration::getPostTypeArgs()));
         $employee->addHooks();
-        $employeeNotifications = new EmployeeNotifications();
+        $employeeNotifications = new EmployeeNotificationFilters();
         $employeeNotifications->addHooks();
 
         $employeeApiManager = new EmployeeApiManager(new EmployeeApiValidator());
@@ -53,5 +54,7 @@ class App
 
         $application = new Application(...array_values(ApplicationConfiguration::getPostTypeArgs()));
         $application->addHooks();
+        $applicationNotifications = new ApplicationNotificationFilters();
+        $applicationNotifications->addHooks();
     }
 }
